Inti

Documentation for Inti.

Inti.AbstractReferenceShape

Inti.ReferenceCube

Inti.ReferenceHyperCube

Inti.ReferenceLine

Inti.ReferenceSimplex

Inti.ReferenceSquare

Inti.ReferenceTetrahedron

Inti.ReferenceTriangle

Inti.ambient_dimension

Inti.geometric_dimension

Inti.standard_basis_vector

Inti.svector

Inti.AbstractReferenceShape — Type

abstract type AbstractReferenceShape

A fixed reference domain/shape. Used mostly for defining more complex shapes as transformations mapping an AbstractReferenceShape to some region of ℜᴹ.

See e.g. ReferenceLine or ReferenceTriangle for some examples of concrete subtypes.

Inti.ReferenceCube — Type

const ReferenceCube = ReferenceHyperCube{3}

Singleton type representing the unit cube [0,1]³.

Inti.ReferenceHyperCube — Type

struct ReferenceHyperCube{N} <: AbstractReferenceShape{N}

Singleton type representing the axis-aligned hypercube in N dimensions with the lower corner at the origin and the upper corner at (1,1,…,1).

Inti.ReferenceLine — Type

const ReferenceLine = ReferenceHyperCube{1}

Singleton type representing the [0,1] segment.

Inti.ReferenceSimplex — Type

struct ReferenceSimplex{N}

Singleton type representing the N-simplex with N+1 vertices (0,...,0),(0,...,0,1),(0,...,0,1,0),(1,0,...,0)

Inti.ReferenceSquare — Type

const ReferenceSquare = ReferenceHyperCube{2}

Singleton type representing the unit square [0,1]².

Inti.ReferenceTetrahedron — Type

struct ReferenceTetrahedron

Singleton type representing the tetrahedron with vertices (0,0,0),(0,0,1),(0,1,0),(1,0,0)

Inti.ReferenceTriangle — Type

struct ReferenceTriangle

Singleton type representing the triangle with vertices (0,0),(1,0),(0,1)

Inti.ambient_dimension — Function

ambient_dimension(x)

Dimension of the ambient space where x lives. For geometrical objects this can differ from its geometric_dimension; for example a triangle in ℝ³ has ambient dimension 3 but geometric dimension 2, while a curve in ℝ³ has ambient dimension 3 but geometric dimension 1.

Inti.geometric_dimension — Function

geometric_dimension(x)

Number of independent coordinates needed to describe x. Lines have geometric dimension 1, triangles have geometric dimension 2, etc.

Inti.standard_basis_vector — Method

standard_basis_vector(k, ::Val{N})

Create an SVector of length N with a 1 in the kth position and zeros elsewhere.

Inti.svector — Method

svector(f,n)

Create an SVector of length n, computing each element as f(i), where i is the index of the element.