References

Inti

Inti.ENTITIES — Constant

const ENTITIES

Global dictionary storing the used entity tags (the value) for a given dimension (the key).

Inti.TAGS — Constant

const TAGS::Dict{Int,Vector{Int}}

Global dictionary storing the used entity tags (the value) for a given dimension (the key).

Inti.AbstractEntity — Type

abstract type AbstractEntity

Entity of geometrical nature. Identifiable throught its (dim,tag) key.

Inti.AbstractMesh — Type

abstract type AbstractMesh{N,T}

An abstract mesh structure in dimension N with primite data of type T (e.g. Float64 for double precision representation).

Concrete subtypes of AbstractMesh should implement ElementIterator for accessing the mesh elements.

See also: LagrangeMesh

Inti.Domain — Type

struct Domain

Represents a physical domain as a union of AbstractEntity objects.

Inti.ElementIterator — Type

struct ElementIterator{E,M}

Iterator for all elements of type E on a mesh of type M.

Besides the methods listed in the iterator iterface of Julia, some functions also require the getindex(iter,i::Int) method for accessing the i-th element directly.

Inti.Fejer — Type

struct Fejer{N}

N-point Fejer's first quadrature rule for integrating a function over [0,1]. Exactly integrates all polynomials of degree ≤ N-1.

using Inti

q = Inti.Fejer(;order=10)

Inti.integrate(cos,q) ≈ sin(1) - sin(0)

# output

true

Inti.Gauss — Type

struct Gauss{D,N} <: ReferenceQuadrature{D}

Tabulated N-point symmetric Gauss quadrature rule for integration over D.

Inti.LagrangeCube — Type

const LagrangeSquare = LagrangeElement{ReferenceSquare}

Inti.LagrangeElement — Type

struct LagrangeElement{D,Np,T} <: ReferenceInterpolant{D,T}

A polynomial p : D → T uniquely defined by its Np values on the Np reference nodes of D.

The return type T should be a vector space (i.e. support addition and multiplication by scalars). For istance, T could be a number or a vector, but not a Tuple.

Inti.LagrangeLine — Type

const LagrangeLine = LagrangeElement{ReferenceLine}

Inti.LagrangeMesh — Type

struct LagrangeMesh{N,T} <: AbstractMesh{N,T}

Data structure representing a generic mesh in an ambient space of dimension N, with data of type T.

The LagrangeMesh can, in principle, store elements of any type. Those are given as a key in the elements dictionary, and the value is a data structure which is capable of reconstructing the elements. For example, for a Lagrange element described by p nodes, the value is a p×Nel matrix of integer, where each columns is a list of tags for the nodes of the element. The nodes are stored in the nodes field.

Inti.LagrangeSquare — Type

const LagrangeSquare = LagrangeElement{ReferenceSquare}

Inti.LagrangeTetrahedron — Type

const LagrangeTetrahedron = LagrangeElement{ReferenceTetrahedron}

Inti.LagrangeTriangle — Type

const LagrangeTriangle = LagrangeElement{ReferenceTriangle}

Inti.Quadrature — Type

struct Quadrature{N,T} <: AbstractVector{QuadratureNode{N,T}}

A collection of QuadratureNodes used to integrate over an AbstractMesh.

Inti.Quadrature — Method

Quadrature(msh::AbstractMesh,etype2qrule::Dict)
Quadrature(msh::AbstractMesh;qorder)

Construct a Quadrature of msh, where for each element type E of msh the reference quadrature q = etype2qrule[E] is used. If an order keyword is passed, a default quadrature of the desired order is used for each element type usig _qrule_for_reference_shape.

For co-dimension 1 elements, the normal vector is computed and stored in the QuadratureNodes.

Inti.QuadratureNode — Type

QuadratureNode{N,T<:Real}

A point in ℝᴺ with a weight for performing numerical integration. A QuadratureNode can optionally store a normal vector.

Inti.ReferenceCube — Type

const ReferenceCube = ReferenceHyperCube{3}

Singleton type representing the unit cube [0,1]³.

Inti.ReferenceHyperCube — Type

struct ReferenceHyperCube{N} <: ReferenceShape{N}

Singleton type representing the axis-aligned hypercube in N dimensions with the lower corner at the origin and the upper corner at (1,1,…,1).

Inti.ReferenceInterpolant — Type

abstract type ReferenceInterpolant{D,T}

Interpolanting function mapping points on the domain D<:ReferenceShape (of singleton type) to a value of type T.

Instances el of ReferenceInterpolant are expected to implement:

el(x̂): evaluate the interpolation scheme at the (reference) coordinate x̂ ∈ D.

jacobian(el,x̂) : evaluate the jacobian matrix of the interpolation at the (reference) coordinate x ∈ D.

Note

For performance reasons, both el(x̂) and jacobian(el,x̂) should take as input a StaticVector and output a static vector or static array.

Inti.ReferenceLine — Type

const ReferenceLine = ReferenceHyperCube{1}

Singleton type representing the [0,1] segment.

Inti.ReferenceQuadrature — Type

abstract type ReferenceQuadrature{D}

A quadrature rule for integrating a function over the domain D <: ReferenceShape.

Calling x,w = q() returns the nodes x, given as SVectors, and weights w, for performing integration over domain(q).

Inti.ReferenceShape — Type

abstract type ReferenceShape

A fixed reference domain/shape. Used mostly for defining more complex shapes as transformations mapping an ReferenceShape to some region of ℜᴹ.

See e.g. ReferenceLine or ReferenceTriangle for some examples of concrete subtypes.

Inti.ReferenceSimplex — Type

struct ReferenceSimplex{N}

Singleton type representing the N-simplex with N+1 vertices (0,...,0),(0,...,0,1),(0,...,0,1,0),(1,0,...,0)

Inti.ReferenceSquare — Type

const ReferenceSquare = ReferenceHyperCube{2}

Singleton type representing the unit square [0,1]².

Inti.ReferenceTetrahedron — Type

struct ReferenceTetrahedron

Singleton type representing the tetrahedron with vertices (0,0,0),(0,0,1),(0,1,0),(1,0,0)

Inti.ReferenceTriangle — Type

struct ReferenceTriangle

Singleton type representing the triangle with vertices (0,0),(1,0),(0,1)

Inti.SubMesh — Type

struct SubMesh{N,T} <: AbstractMesh{N,T}

Create a view of a parent LagrangeMesh over a given domain.

A submesh implements the interface for AbstractMesh; therefore you can iterate over elements of the submesh just like you would with a mesh.

Inti.TensorProductQuadrature — Type

TensorProductQuadrature{N,Q}

A tensor-product of one-dimension quadrature rules. Integrates over [0,1]^N.

Examples

qx = Fejer(10)
qy = TrapezoidalOpen(15)
q  = TensorProductQuadrature(qx,qy)

Base.:== — Method

==(e1::AbstractEntity,e2::AbstractEntity)

Two entities are considered equal geometric_dimension(e1)==geometric_dimension(e2) and tag(e1)=tag(e2).

Notice that this implies dim and tag of an entity should uniquely define it, and therefore global variables like TAGS are needed to make sure newly created AbstractEntity have a new (dim,tag) identifier.

Base.iterate — Function

iterate(Ω::Domain)

Iterating over a domain means iterating over its entities.

Inti._get_gauss_qcoords_and_qweights — Method

_get_gauss_and_qweights(R::Type{<:ReferenceShape{D}}, N) where D

Returns the N-point symmetric gaussian qnodes and qweights (x, w) for integration over R.

Inti._integration_measure — Method

_integration_measure(J::AbstractMatrix)

Given the Jacobian matrix J of a transformation f : ℝᴹ → ℝᴺ, compute the integration measure √det(JᵀJ).

Inti._normal — Method

_normal(jac::SMatrix{M,N})

Given a an M by N matrix representing the jacobian of a codimension one object, compute the normal vector.

Inti._qrule_for_reference_shape — Method

_qrule_for_reference_shape(ref,order)

Given a reference shape and a desired quadrature order, return an appropiate quadrature rule.

Inti.ambient_dimension — Function

ambient_dimension(x)

Dimension of the ambient space where x lives. For geometrical objects this can differ from its geometric_dimension; for example a triangle in ℝ³ has ambient dimension 3 but geometric dimension 2, while a curve in ℝ³ has ambient dimension 3 but geometric dimension 1.

Inti.boundary — Method

boundary(e::AbstractEntity)

A vector of entities of dimension dim-1 that form the boundary of e.

Inti.clear_entities! — Method

clear_entities!()

Empty the global variables used to keep track of the various entities created.

See also: ENTITIES, TAGS

Inti.degree — Method

degree(el::LagrangeElement)
degree(el::Type{<:LagrangeElement})

The polynomial degree el.

Inti.dom2elt — Method

dom2elt(m::LagrangeMesh,Ω,E)

Compute the element indices idxs of the elements of type E composing Ω, so that elements(m)[idxs] gives all the elements of type E meshing Ω.

Inti.domain — Function

domain(f)

Given a function-like object f: Ω → R, return Ω.

Inti.domain — Method

domain(q::ReferenceQuadrature)

The domain of integratino for quadrature rule q.

Inti.element_types — Method

element_types(msh::AbstractMesh)

Return the element types present in the msh.

Inti.elements — Method

elements(msh::AbstractMesh,E::DataType)

Return an iterator for all elements of type E on a mesh msh.

Inti.entities — Method

entities(Ω::Domain)

Return all entities making up a domain.

Inti.external_boundary — Method

external_boundary(Ω::Domain)

Return the external boundaries inside a domain. These are entities in the skeleton of Ω which are not in the internal boundaries of Ω.

See also: internal_boundary, skeleton.

Inti.geometric_dimension — Function

geometric_dimension(x)

NNumber of degrees of freedom necessary to locally represent the geometrical object. For example, lines have geometric dimension of 1 (whether in ℝ² or in ℝ³), while surfaces have geometric dimension of 2.

Inti.get_gmsh_extension — Method

get_gmsh_extension()

Get the Gmsh extension, if available.

Inti.get_makie_extension — Method

get_makie_extension()

Get the Makie extension, if available.

Inti.get_vtk_extension — Method

get_vtk_extension()

Get the VTK extension, if available.

Inti.global_add_entity! — Method

global_add_entity!(ent::AbstractEntity)

Add ent to the global dictionary ENTITIES and update TAGS with its (dim,tag) key. This function should be called by the inner constructor of every AbstractEntity.

Inti.image — Function

image(f)

Given a function-like object f: Ω → R, return f(Ω).

Inti.integrate — Method

integrate(f,quad::Quadrature)

Compute ∑ᵢ f(qᵢ)wᵢ, where the qᵢ are the quadrature nodes of quad, and wᵢ are the quadrature weights.

Note that you must define f(::QuadratureNode): use q.coords and q.normal if you need to access the coordinate or normal vector at que quadrature node.

Inti.integrate — Method

integrate(f,q::ReferenceQuadrature)
integrate(f,x,w)

Integrate the function f using the quadrature rule q. This is simply sum(f.(x) .* w), where x and w are the quadrature nodes and weights, respectively.

The function f should take an SVector as input.

Inti.interface_method — Method

interface_method(x)

A method of an abstract type for which concrete subtypes are expected to provide an implementation.

Inti.internal_boundary — Method

internal_boundary(Ω::Domain)

Return the internal boundaries of a Domain. These are entities in skeleton(Ω) which appear at least twice as a boundary of entities in Ω.

Inti.jacobian — Method

jacobian(f,x)

Given a (possibly vector-valued) functor f : 𝐑ᵐ → 𝐅ⁿ, return the n × m matrix Aᵢⱼ = ∂fᵢ/∂xⱼ.Both x and f(x) are expected to be of SVector type.

Inti.key — Method

key(e::AbstractEntity)

The (dim,tag) pair used as a key to identify geometrical entities.

Inti.new_tag — Method

new_tag(dim)

Generate a unique tag for an AbstractEntity of dimension dim.

The implementation consists of adding one to the maximum value of TAGS[dim]

See also: TAGS.

Inti.order — Method

order(q::ReferenceQuadrature)

A quadrature of order p integrates all polynomials of degree ≤ p.

Inti.qcoords — Method

qcoords(q)

Return the coordinate of the quadrature nodes associated with q.

Inti.qweights — Method

qweights(q)

Return the quadrature weights associated with q.

Inti.reference_nodes — Method

reference_nodes(el::LagrangeElement)
reference_nodes(::Type{<:LagrangeElement})

Return the reference nodes on domain(el) used for the polynomial interpolation. The function values on these nodes completely determines the interpolating polynomial.

We use the same convention as gmsh for defining the reference nodes and their order (see node ordering on gmsh documentation).

Inti.return_type — Method

return_type(f[,args...])

The type returned by f(args...), where args is a tuple of types. Falls back to Base.promote_op by default.

A functors of type T with a knonw return type should extend return_type(::T,args...) to avoid relying on promote_op.

Inti.skeleton — Method

skeleton(Ω::Domain)

Return all the boundaries of the domain, i.e. the domain's skeleton.

Inti.standard_basis_vector — Method

standard_basis_vector(k, ::Val{N})

Create an SVector of length N with a 1 in the kth position and zeros elsewhere.

Inti.svector — Method

svector(f,n)

Create an SVector of length n, computing each element as f(i), where i is the index of the element.

Inti.tag — Method

tag(::AbstractEntity)

Integer tag used to idetify geometrical entities.

IntiGmshExtension

IntiGmshExt.GmshEntity — Type

struct GmshEntity <: AbstractEntity

Concrete type of AbstractEntity generated using the gmsh API.

IntiGmshExt._domain_to_mesh! — Method

_domain_to_mesh!(msh, Ω::Domain)

Recursively populate the dictionaries etype2mat and ent2tag in msh for the entities in Ω. After all entities have been processed, the function recurses on the skeleton of Ω.

IntiGmshExt._ent_to_mesh! — Method

_ent_to_mesh!(elements, ent2tag, ω::ElementaryEntity)

For each element type used to mesh ω:

push into elements::Dict the pair etype=>ntags;

push into ent2tag::Dict the pair etype=>etags;

where:

etype::DataType determines the type of the element (see _type_tag_to_etype);

ntags::Matrix{Int} gives the indices of the nodes defining those elements;

etags::Vector{Int} gives the indices of those elements in elements.

IntiGmshExt._fill_entity_boundary! — Method

_fill_entity_boundary!

Use the gmsh API to add the boundary of an ElementaryEntity.

This is a helper function, and should not be called by itself.

IntiGmshExt._import_domain! — Function

_import_domain!(Ω::Domain,[model;dim=3])

Like import_domain, but appends entities to Ω instead of creating a new domain.

Note

This function assumes that gmsh has been initialized, and does not handle its finalization.

IntiGmshExt._import_mesh! — Method

_import_mesh!(msh,Ω)

Similar to import_mesh, but append information to msh instead of creating a new mesh.

Danger

This function assumes that gmsh has been initialized, and does not handle its finalization.

IntiGmshExt._type_tag_to_etype — Method

_type_tag_to_etype(tag)

Mapping of gmsh element types, encoded as an integer, to the internal equivalent of those.

IntiGmshExt.import_domain — Function

import_domain([model;dim=3])

Construct a Domain from the gmsh model with all entities of dimension dim; by defaul the current gmsh model is used.

Note

This function assumes that gmsh has been initialized, and does not handle its finalization.

IntiGmshExt.import_mesh — Method

import_mesh(Ω;[dim=3])

Create a LagrangeMesh for the entities in Ω. Passing dim=2 will create a two-dimensional mesh by projecting the original mesh onto the x,y plane.

Danger

This function assumes that gmsh has been initialized, and does not handle its finalization.

IntiGmshExt.model_summary — Function

model_summary([model])

Print a summary of the gmsh model to the console; by defaul the current model is used.

IntiGmshExt.read_geo — Method

read_geo(fname::String;dim=3)

Read a .geo file and generate a Domain with all entities of dimension dim.

Danger

This function assumes that gmsh has been initialized, and does not handle its finalization.

IntiMakieExtension

IntiVTKExtension

IntiVTKExt.etype_to_vtk_cell_type — Constant

const etype_to_vtk_cell_type

Dictionary mapping internal element types to a tuple containing: - the corresponding WriteVTK cell types (following the convention chosen by VTK, see below); - the indices in the elements column that defines the element. This is because we want to support the export of more than just the flat elements available in the VTK specification, hence which may require a conversion of some sort.

See VTK specification [Fig. 2] on http://www.vtk.org/VTK/img/file-formats.pdf

VTK_VERTEX (=1)

VTKPOLYVERTEX (=2)

VTK_LINE (=3)

VTKPOLYLINE (=4)

VTK_TRIANGLE (=5)

VTKTRIANGLESTRIP (=6)

VTK_POLYGON (=7)

VTK_PIXEL (=8)

VTK_QUAD (=9)

VTK_TETRA (=10)

VTK_VOXEL (=11)

VTK_HEXAHEDRON (=12)

VTK_WEDGE (=13)

VTK_PYRAMID (=14)

IntiVTKExt._vtk_cells — Function

_vtk_cells(mesh::LagrangeMesh, E::DataType)
_vtk_cells(mesh::LagrangeMesh, Ω::Domain)

Creates the vector of all elements contained in the mesh in the format required by WriteVTK for a particular element type E<:AbstractElement or a Domain instance.

IntiVTKExt._vtk_points — Method

_vtk_points(mesh::LagrangeMesh)

Creates the matrix of nodes in the format required by WriteVTK.

IntiVTKExt.mesh_file — Method

mesh_file(mesh::LagrangeMesh[, Ω::Domain], name::String)

Creates a VTK file (.vtu) with name name containing the mesh information. It is possible to export only a Domain (i.e. only a part of the mesh).

Note that all the heavy lifting is done by the package WriteVTK. We refer to its documentation for more information.

Warning: the output is not an actual file (on disk). To save it, simply write:

mesh_file(mesh, "my_mesh") |> vtk_save

(Noting that you will need to be using WriteVTK to do so)

To add some data (scalar or vector values at the mesh nodes or mesh cells) for visualization purposes, you can do for instance:

vtkfile = mesh_file(mesh, name)
vtkfile["my_point_data", VTKPointData()] = pdata
vtkfile["my_cell_data", VTKCellData()] = cdata
vtk_save(vtkfile)

It is possible also to export a partition Ωs::Vector{Domain} using Multiblock files (.vtm), for instance like so

vtmfile = vtk_multiblock(name)
for (Ω, pdata) in zip(Ωs, pdatas)
    vtkfile = mesh_file(mesh, Ω, name)
    vtkfile["my_point_data", VTKPointData()] = pdata
end
vtk_save(vtmfile)

To save a sequence of solutions (time steps, iterations), simply append the number of the element to the file name. Paraview will recognize the sequence automatically.