Inti

Documentation for Inti.

Inti.Fejer

Inti.Gauss

Inti.LagrangeCube

Inti.LagrangeElement

Inti.LagrangeLine

Inti.LagrangeSquare

Inti.LagrangeTetrahedron

Inti.LagrangeTriangle

Inti.ReferenceCube

Inti.ReferenceHyperCube

Inti.ReferenceInterpolant

Inti.ReferenceLine

Inti.ReferenceQuadrature

Inti.ReferenceShape

Inti.ReferenceSimplex

Inti.ReferenceSquare

Inti.ReferenceTetrahedron

Inti.ReferenceTriangle

Inti.TensorProductQuadrature

Inti._get_gauss_qcoords_and_qweights

Inti.ambient_dimension

Inti.degree

Inti.domain

Inti.domain

Inti.geometric_dimension

Inti.image

Inti.integrate

Inti.interface_method

Inti.jacobian

Inti.order

Inti.qcoords

Inti.qrule_for_reference_shape

Inti.qweights

Inti.reference_nodes

Inti.return_type

Inti.standard_basis_vector

Inti.svector

Inti.Fejer — Type

struct Fejer{N}

N-point Fejer's first quadrature rule for integrating a function over [0,1]. Exactly integrates all polynomials of degree ≤ N-1.

using Inti

q = Inti.Fejer(;order=10)

Inti.integrate(cos,q) ≈ sin(1) - sin(0)

# output

true

Inti.Gauss — Type

struct Gauss{D,N} <: ReferenceQuadrature{D}

Tabulated N-point symmetric Gauss quadrature rule for integration over D.

Inti.LagrangeCube — Type

const LagrangeSquare = LagrangeElement{ReferenceSquare}

Inti.LagrangeElement — Type

struct LagrangeElement{D,Np,T} <: ReferenceInterpolant{D,T}

A polynomial p : D → T uniquely defined by its Np values on the Np reference nodes of D.

The return type T should be a vector space (i.e. support addition and multiplication by scalars). For istance, T could be a number or a vector, but not a Tuple.

Inti.LagrangeLine — Type

const LagrangeLine = LagrangeElement{ReferenceLine}

Inti.LagrangeSquare — Type

const LagrangeSquare = LagrangeElement{ReferenceSquare}

Inti.LagrangeTetrahedron — Type

const LagrangeTetrahedron = LagrangeElement{ReferenceTetrahedron}

Inti.LagrangeTriangle — Type

const LagrangeTriangle = LagrangeElement{ReferenceTriangle}

Inti.ReferenceCube — Type

const ReferenceCube = ReferenceHyperCube{3}

Singleton type representing the unit cube [0,1]³.

Inti.ReferenceHyperCube — Type

struct ReferenceHyperCube{N} <: ReferenceShape{N}

Singleton type representing the axis-aligned hypercube in N dimensions with the lower corner at the origin and the upper corner at (1,1,…,1).

Inti.ReferenceInterpolant — Type

abstract type ReferenceInterpolant{D,T}

Interpolanting function mapping points on the domain D<:ReferenceShape (of singleton type) to a value of type T.

Instances el of ReferenceInterpolant are expected to implement:

el(x̂): evaluate the interpolation scheme at the (reference) coordinate x̂ ∈ D.

jacobian(el,x̂) : evaluate the jacobian matrix of the interpolation at the (reference) coordinate x ∈ D.

Note

For performance reasons, both el(x̂) and jacobian(el,x̂) should take as input a StaticVector and output a static vector or static array.

Inti.ReferenceLine — Type

const ReferenceLine = ReferenceHyperCube{1}

Singleton type representing the [0,1] segment.

Inti.ReferenceQuadrature — Type

abstract type ReferenceQuadrature{D}

A quadrature rule for integrating a function over the domain D <: ReferenceShape.

Calling x,w = q() returns the nodes x, given as SVectors, and weights w, for performing integration over domain(q).

Inti.ReferenceShape — Type

abstract type ReferenceShape

A fixed reference domain/shape. Used mostly for defining more complex shapes as transformations mapping an ReferenceShape to some region of ℜᴹ.

See e.g. ReferenceLine or ReferenceTriangle for some examples of concrete subtypes.

Inti.ReferenceSimplex — Type

struct ReferenceSimplex{N}

Singleton type representing the N-simplex with N+1 vertices (0,...,0),(0,...,0,1),(0,...,0,1,0),(1,0,...,0)

Inti.ReferenceSquare — Type

const ReferenceSquare = ReferenceHyperCube{2}

Singleton type representing the unit square [0,1]².

Inti.ReferenceTetrahedron — Type

struct ReferenceTetrahedron

Singleton type representing the tetrahedron with vertices (0,0,0),(0,0,1),(0,1,0),(1,0,0)

Inti.ReferenceTriangle — Type

struct ReferenceTriangle

Singleton type representing the triangle with vertices (0,0),(1,0),(0,1)

Inti.TensorProductQuadrature — Type

TensorProductQuadrature{N,Q}

A tensor-product of one-dimension quadrature rules. Integrates over [0,1]^N.

Examples

qx = Fejer(10)
qy = TrapezoidalOpen(15)
q  = TensorProductQuadrature(qx,qy)

Inti._get_gauss_qcoords_and_qweights — Method

_get_gauss_and_qweights(R::Type{<:ReferenceShape{D}}, N) where D

Returns the N-point symmetric gaussian qnodes and qweights (x, w) for integration over R.

Inti.ambient_dimension — Function

ambient_dimension(x)

Dimension of the ambient space where x lives. For geometrical objects this can differ from its geometric_dimension; for example a triangle in ℝ³ has ambient dimension 3 but geometric dimension 2, while a curve in ℝ³ has ambient dimension 3 but geometric dimension 1.

Inti.degree — Method

degree(el::LagrangeElement)
degree(el::Type{<:LagrangeElement})

The polynomial degree el.

Inti.domain — Function

domain(f)

Given a function-like object f: Ω → R, return Ω.

Inti.domain — Method

domain(q::ReferenceQuadrature)

The domain of integratino for quadrature rule q.

Inti.geometric_dimension — Function

geometric_dimension(x)

NNumber of degrees of freedom necessary to locally represent the geometrical object. For example, lines have geometric dimension of 1 (whether in ℝ² or in ℝ³), while surfaces have geometric dimension of 2.

Inti.image — Function

image(f)

Given a function-like object f: Ω → R, return f(Ω).

Inti.integrate — Method

integrate(f,q::ReferenceQuadrature)
integrate(f,x,w)

Integrate the function f using the quadrature rule q. This is simply sum(f.(x) .* w), where x and w are the quadrature nodes and weights, respectively.

The function f should take an SVector as input.

Inti.interface_method — Method

interface_method(x)

A method of an abstract type for which concrete subtypes are expected to provide an implementation.

Inti.jacobian — Method

jacobian(f,x)

Given a (possibly vector-valued) functor f : 𝐑ᵐ → 𝐅ⁿ, return the n × m matrix Aᵢⱼ = ∂fᵢ/∂xⱼ.Both x and f(x) are expected to be of SVector type.

Inti.order — Method

order(q::ReferenceQuadrature)

A quadrature of order p integrates all polynomials of degree ≤ p.

Inti.qcoords — Method

qcoords(q)

Return the coordinate of the quadrature nodes associated with q.

Inti.qrule_for_reference_shape — Method

qrule_for_reference_shape(ref,order)

Given a reference shape and a desired quadrature order, return an appropiate quadrature rule.

Inti.qweights — Method

qweights(q)

Return the quadrature weights associated with q.

Inti.reference_nodes — Method

reference_nodes(el::LagrangeElement)
reference_nodes(::Type{<:LagrangeElement})

Return the reference nodes on domain(el) used for the polynomial interpolation. The function values on these nodes completely determines the interpolating polynomial.

We use the same convention as gmsh for defining the reference nodes and their order (see node ordering on gmsh documentation).

Inti.return_type — Method

return_type(f[,args...])

The type returned by f(args...), where args is a tuple of types. Falls back to Base.promote_op by default.

A functors of type T with a knonw return type should extend return_type(::T,args...) to avoid relying on promote_op.

Inti.standard_basis_vector — Method

standard_basis_vector(k, ::Val{N})

Create an SVector of length N with a 1 in the kth position and zeros elsewhere.

Inti.svector — Method

svector(f,n)

Create an SVector of length n, computing each element as f(i), where i is the index of the element.