References

Inti.ENTITIES — Constant

const ENTITIES

Dictionary mapping EntityKey to GeometricEntity. Contains all entities created in a given session.

Inti.SAME_POINT_TOLERANCE — Constant

SAME_POINTS_TOLERANCE

Two points x and y are considerd the same if norm(x-y) ≤ SAME_POINT_TOLERANCE.

Inti.AbstractKernel — Type

abstract type AbstractKernel{T}

A kernel functions K with the signature K(target,source)::T.

See also: GenericKernel, SingleLayerKernel, DoubleLayerKernel, AdjointDoubleLayerKernel, HyperSingularKernel

Inti.AbstractMesh — Type

abstract type AbstractMesh{N,T}

An abstract mesh structure in dimension N with primite data of type T (e.g. Float64 for double precision representation).

Concrete subtypes of AbstractMesh should implement ElementIterator for accessing the mesh elements.

See also: LagrangeMesh

Inti.AbstractPDE — Type

abstract type AbstractPDE{N}

A partial differential equation in dimension N. AbstractPDE types are used to define AbstractPDEKernels.

Inti.AbstractPDEKernel — Type

abstract type AbstractPDEKernel{T,Op} <: AbstractKernel{T}

An AbstractKernel with an associated pde::Op field.

Inti.AdjointDoubleLayerKernel — Type

struct AdjointDoubleLayerKernel{T,Op} <: AbstractPDEKernel{T,Op}

Given an operator Op, construct its free-space adjoint double-layer kernel. This corresponds to the transpose(γ₁,ₓ[G]), where G is the SingleLayerKernel. For operators such as Laplace or Helmholtz, this is simply the normal derivative of the fundamental solution respect to the target variable.

Inti.DimParameters — Type

struct DimParameters

Parameters associated with the density interpolation method used in bdim_correction.

Inti.Domain — Type

struct Domain

A set of EntityKeys with the same geometric dimension used to represent a physical domain.

Inti.Domain — Method

Domain([f::Function,] ents)

Create a domain from a set of EntityKeys. Optionally, a filter function f can be passed to filter the entities.

Note that all entities in a domain must have the same geometric dimension.

Inti.DoubleLayerKernel — Type

struct DoubleLayerKernel{T,Op} <: AbstractPDEKernel{T,Op}

Given an operator Op, construct its free-space double-layer kernel. This corresponds to the γ₁ trace of the SingleLayerKernel. For operators such as Laplace or Helmholtz, this is simply the normal derivative of the fundamental solution respect to the source variable.

Inti.ElementIterator — Type

struct ElementIterator{E,M}

Iterator for all elements of type E on a mesh of type M.

Besides the methods listed in the iterator iterface of Julia, some functions also require the getindex(iter,i::Int) method for accessing the i-th element directly.

Inti.EntityKey — Type

EntityKey

Used to represent the key of a GeometricEntity, comprised of a dim and a tag field, where dim is the geometrical dimension of the entity, and tag is a unique integer identifying the entity.

The sign of the tag field is used to distinguish the orientation of the entity, and is ignored when comparing two EntityKeys for equality.

Inti.Fejer — Type

struct Fejer{N}

N-point Fejer's first quadrature rule for integrating a function over [0,1]. Exactly integrates all polynomials of degree ≤ N-1.

using Inti

q = Inti.Fejer(;order=10)

Inti.integrate(cos,q) ≈ sin(1) - sin(0)

# output

true

Inti.Gauss — Type

struct Gauss{D,N} <: ReferenceQuadrature{D}

Tabulated N-point symmetric Gauss quadrature rule for integration over D.

Inti.GenericKernel — Type

struct GenericKernel{T,F} <: AbstractKernel{T}

An AbstractKernel with kernel of type F.

Inti.GeometricEntity — Type

struct GeometricEntity

Used to represent geometrical objects such as lines, surfaces, and volumes.

Geometrical entities are stored in a global ENTITIES dictionary mapping EntityKey to the corresponding GeometricEntity.

A GeometricEntitymay also have a push_forward field associated with it, which provides a map from a reference domain to the entity itself.

Inti.Helmholtz — Type

struct Helmholtz{N,T}

Helmholtz equation in N dimensions: Δu + k²u = 0.

Inti.HyperSingularKernel — Type

struct HyperSingularKernel{T,Op} <: AbstractPDEKernel{T,Op}

Given an operator Op, construct its free-space hypersingular kernel. This corresponds to the transpose(γ₁,ₓγ₁[G]), where G is the SingleLayerKernel. For operators such as Laplace or Helmholtz, this is simply the normal derivative of the fundamental solution respect to the target variable of the DoubleLayerKernel.

Inti.IntegralOperator — Type

struct IntegralOperator{T} <: AbstractMatrix{T}

A discrete linear integral operator given by

\[I[u](x) = \int_{\Gamma\_s} K(x,y)u(y) ds_y, x \in \Gamma_{t}\]

where $\Gamma_s$ and $\Gamma_t$ are the source and target domains, respectively.

Inti.IntegralPotential — Type

struct IntegralPotential

Represent a potential given by a kernel and a quadrature over which integration is performed.

IntegralPotentials are created using IntegralPotential(kernel, quadrature).

Evaluating an integral potential requires a density σ (defined over the quadrature nodes of the source mesh) and a point x at which to evaluate the integral

\[\int_{\Gamma} K(oldsymbol{x},oldsymbol{y})\sigma(y) ds_y, x \not \in \Gamma\]

Assuming 𝒮 is an integral potential and σ is a vector of values defined on quadrature, calling 𝒮[σ] creates an anonymous function that can be evaluated at any point x.

Inti.LagrangeCube — Type

const LagrangeSquare = LagrangeElement{ReferenceSquare}

Inti.LagrangeElement — Type

struct LagrangeElement{D,Np,T} <: ReferenceInterpolant{D,T}

A polynomial p : D → T uniquely defined by its Np values on the Np reference nodes of D.

The return type T should be a vector space (i.e. support addition and multiplication by scalars). For istance, T could be a number or a vector, but not a Tuple.

Inti.LagrangeLine — Type

const LagrangeLine = LagrangeElement{ReferenceLine}

Inti.LagrangeMesh — Type

struct LagrangeMesh{N,T} <: AbstractMesh{N,T}

Data structure representing a generic mesh in an ambient space of dimension N, with data of type T.

The LagrangeMesh can, in principle, store elements of any type. Those are given as a key in the elements dictionary, and the value is a data structure which is capable of reconstructing the elements. For example, for a Lagrange element described by p nodes, the value is a p×Nel matrix of integer, where each columns is a list of tags for the nodes of the element. The nodes are stored in the nodes field.

Inti.LagrangeSquare — Type

const LagrangeSquare = LagrangeElement{ReferenceSquare}

Inti.LagrangeTetrahedron — Type

const LagrangeTetrahedron = LagrangeElement{ReferenceTetrahedron}

Inti.LagrangeTriangle — Type

const LagrangeTriangle = LagrangeElement{ReferenceTriangle}

Inti.Laplace — Type

struct Laplace{N}

Laplace equation in N dimension: Δu = 0.

Inti.MultiIndex — Type

MultiIndex{N}

Wrapper around NTuple{N,Int} mimicking a multi-index in ℤ₀ᴺ.

Inti.Quadrature — Type

struct Quadrature{N,T} <: AbstractVector{QuadratureNode{N,T}}

A collection of QuadratureNodes used to integrate over an AbstractMesh.

Inti.Quadrature — Method

Quadrature(msh::AbstractMesh, etype2qrule::Dict)
Quadrature(msh::AbstractMesh; qorder)

Construct a Quadrature for msh, where for each element type E in msh the reference quadrature q = etype2qrule[E] is used. If an order keyword is passed, a default quadrature of the desired order is used for each element type usig _qrule_for_reference_shape.

For co-dimension one elements, the normal vector is also computed and stored in the QuadratureNodes.

Inti.QuadratureNode — Type

QuadratureNode{N,T<:Real}

A point in ℝᴺ with a weight for performing numerical integration. A QuadratureNode can optionally store a normal vector.

Inti.ReferenceCube — Type

const ReferenceCube = ReferenceHyperCube{3}

Singleton type representing the unit cube [0,1]³.

Inti.ReferenceHyperCube — Type

struct ReferenceHyperCube{N} <: ReferenceShape{N}

Singleton type representing the axis-aligned hypercube in N dimensions with the lower corner at the origin and the upper corner at (1,1,…,1).

Inti.ReferenceInterpolant — Type

abstract type ReferenceInterpolant{D,T}

Interpolanting function mapping points on the domain D<:ReferenceShape (of singleton type) to a value of type T.

Instances el of ReferenceInterpolant are expected to implement:

el(x̂): evaluate the interpolation scheme at the (reference) coordinate x̂ ∈ D.

jacobian(el,x̂) : evaluate the jacobian matrix of the interpolation at the (reference) coordinate x ∈ D.

Note

For performance reasons, both el(x̂) and jacobian(el,x̂) should take as input a StaticVector and output a static vector or static array.

Inti.ReferenceLine — Type

const ReferenceLine = ReferenceHyperCube{1}

Singleton type representing the [0,1] segment.

Inti.ReferenceQuadrature — Type

abstract type ReferenceQuadrature{D}

A quadrature rule for integrating a function over the domain D <: ReferenceShape.

Calling x,w = q() returns the nodes x, given as SVectors, and weights w, for performing integration over domain(q).

Inti.ReferenceShape — Type

abstract type ReferenceShape

A fixed reference domain/shape. Used mostly for defining more complex shapes as transformations mapping an ReferenceShape to some region of ℜᴹ.

See e.g. ReferenceLine or ReferenceTriangle for some examples of concrete subtypes.

Inti.ReferenceSimplex — Type

struct ReferenceSimplex{N}

Singleton type representing the N-simplex with N+1 vertices (0,...,0),(0,...,0,1),(0,...,0,1,0),(1,0,...,0)

Inti.ReferenceSquare — Type

const ReferenceSquare = ReferenceHyperCube{2}

Singleton type representing the unit square [0,1]².

Inti.ReferenceTetrahedron — Type

struct ReferenceTetrahedron

Singleton type representing the tetrahedron with vertices (0,0,0),(0,0,1),(0,1,0),(1,0,0)

Inti.ReferenceTriangle — Type

struct ReferenceTriangle

Singleton type representing the triangle with vertices (0,0),(1,0),(0,1)

Inti.SingleLayerKernel — Type

struct SingleLayerKernel{T,Op} <: AbstractPDEKernel{T,Op}

The free-space single-layer kernel (i.e. the fundamental solution) of an OP <: AbstractPDE.

Inti.SubMesh — Type

struct SubMesh{N,T} <: AbstractMesh{N,T}

View into a parent mesh over a given domain.

A submesh implements the interface for AbstractMesh; therefore you can iterate over elements of the submesh just like you would with a mesh.

Construct SubMeshs using view(parent,Ω::Domain).

Inti.TensorProductQuadrature — Type

TensorProductQuadrature{N,Q}

A tensor-product of one-dimension quadrature rules. Integrates over [0,1]^N.

Examples

qx = Fejer(10)
qy = TrapezoidalOpen(15)
q  = TensorProductQuadrature(qx,qy)

Inti.VioreanuRokhlin — Type

struct VioreanuRokhlin{D,N} <: ReferenceQuadrature{D}

Tabulated N-point Vioreanu-Rokhlin quadrature rule for integration over D.

Base.iterate — Function

iterate(Ω::Domain)

Iterating over a domain means iterating over its entities.

Inti._copyto! — Method

_copyto!(target,source)

Defaults to Base.copyto!, but includes some specialized methods to copy from a Matrix of SMatrix to a Matrix of Numbers and viceversa.

Inti._get_gauss_qcoords_and_qweights — Method

_get_gauss_and_qweights(R::Type{<:ReferenceShape{D}}, N) where D

Returns the N-point symmetric gaussian qnodes and qweights (x, w) for integration over R.

Inti._get_vioreanurokhlin_qcoords_and_qweights — Method

_get_vioreanurokhlin_qcoords_and_qweights(R::Type{<:ReferenceShape{D}}, N) where D

Returns the N-point Vioreanu-Rokhlin qnodes and qweights (x, w) for integration over R.

Inti._green_multiplier — Method

_green_multiplier(x, quad)

Helper function to help determine the constant σ in the Green identity S[γ₁u](x)

D[γ₀u](x) + σ*u(x) = 0. This can be used as a predicate to determine whether a

point is inside a domain or not.

Inti._integration_measure — Method

_integration_measure(J::AbstractMatrix)

Given the Jacobian matrix J of a transformation f : ℝᴹ → ℝᴺ, compute the integration measure √det(JᵀJ).

Inti._normal — Method

_normal(jac::SMatrix{M,N})

Given a an M by N matrix representing the jacobian of a codimension one object, compute the normal vector.

Inti._qrule_for_reference_shape — Method

_qrule_for_reference_shape(ref,order)

Given a reference shape and a desired quadrature order, return an appropiate quadrature rule.

Inti.adj_double_layer_hypersingular — Method

adj_double_layer_hypersingular(; pde, target, source, compression,
correction)

Similar to single_double_layer, but for the adjoint double-layer and hypersingular operators. See the documentation of [single_double_layer] for a description of the arguments.

Inti.ambient_dimension — Function

ambient_dimension(x)

Dimension of the ambient space where x lives. For geometrical objects this can differ from its geometric_dimension; for example a triangle in ℝ³ has ambient dimension 3 but geometric dimension 2, while a curve in ℝ³ has ambient dimension 3 but geometric dimension 1.

Inti.assemble_fmm — Method

assemble_fmm(iop; atol)

Set up a 2D or 3D FMM for evaluating the discretized integral operator iop associated with the pde. In 2D the FMMLIB2D library is used while in 3D FMM3D is used.

FMMLIB2D

FMMLIB2D does no checking for if the targets and sources coincide, and will return Inf values if iop.target !== iop.source, but there is a point x ∈ iop.target such that x ∈ iop.source.

Inti.assemble_hmatrix — Method

assemble_hmatrix(iop[; atol, rank, rtol, eta])

Assemble an H-matrix representation of the discretized integral operator iop using the HMatrices.jl library.

See the assemble_hmatrix function from HMatrices.jl for more details on the keyword arguments.

Inti.assemble_matrix — Method

assemble_matrix(iop::IntegralOperator; threads = true)

Assemble the dense matrix representation of an IntegralOperator.

Inti.bdim_correction — Method

bdim_correction(pde,X,Y,S,D[;parameters,derivative,maxdist])

Given a pde and a (possibly innacurate) discretizations of its single and double-layer operators S and D (taking a vector of values on Y and returning a vector on of values on X), compute corrections δS and δD such that S + δS and D + δD are more accurate approximations of the underlying single- and double-layer integral operators.

Requirements:

pde must be an AbstractPDE

Y must be a Quadrature object of a closed surface

X is either on, inside, or outside Y

The following optional keyword arguments are available:

parameters::DimParameters: parameters associated with the density interpolation method

derivative: if true, compute the correction to the adjoint double-layer and hypersingular operators instead. In this case, S and D should be replaced by a (possibly innacurate) discretization of adjoint double-layer and hypersingular operators, respectively.

maxdist: distance beyond which interactions are considered sufficiently far so that no correction is needed. This is used to determine a threshold for nearly-singular corrections when X and Y are different surfaces. When X === Y, this is not needed.

green_multiplier: the value of μ in the Green identity S\[γ₁u\](x) - D\[γ₀u\](x) + μ*u(x) = 0 for x ∈ X. When a scalar value is passed, it is taken to be the same for all values of x. If nothing, the value is inferred using _green_multiplier and the quadrature Y (assumes X is either inside, on, or outside Y).

Inti.boundary — Method

boundary(Ω::Domain)

Return the external boundaries of a domain.

See also: external_boundary, internal_boundary, skeleton.

Inti.coords — Method

coords(q)

Return the spatial coordinates of q.

Inti.degree — Method

degree(el::LagrangeElement)
degree(el::Type{<:LagrangeElement})

The polynomial degree el.

Inti.dom2elt — Method

dom2elt(m::LagrangeMesh,Ω,E)::Vector{Int}

Compute the element indices idxs of the elements of type E composing Ω.

Inti.dom2qtags — Method

dom2qtags(Q::Quadrature, dom::Domain)

Given a domain, return the indices of the quadratures nodes in Q associated to its quadrature.

Inti.domain — Function

domain(f)

Given a function-like object f: Ω → R, return Ω.

Inti.domain — Method

domain(msh::LagrangeMesh)

Set of all entities covered by the mesh.

Inti.domain — Method

domain(Q::Quadrature)

The Domain over which Q performs integration.

Inti.domain — Method

domain(q::ReferenceQuadrature)

The domain of integratino for quadrature rule q.

Inti.element_types — Method

element_types(msh::AbstractMesh)

Return the element types present in the msh.

Inti.elements — Method

elements(msh::AbstractMesh,E::DataType)

Return an iterator for all elements of type E on a mesh msh.

Inti.entities — Method

entities(Ω::Domain)

Return all entities making up a domain.

Inti.etype_to_nearest_points — Method

etype_to_nearest_points(X,Y::Quadrature; maxdist)

For each element el in Y.mesh, return a list with the indices of all points in X for which el is the nearest element. Ignore indices for which the distance exceeds maxdist.

Inti.external_boundary — Method

external_boundary(Ω::Domain)

Return the external boundaries inside a domain. These are entities in the skeleton of Ω which are not in the internal boundaries of Ω.

See also: internal_boundary, skeleton.

Inti.fibonnaci_points_sphere — Method

fibonnaci_points_sphere(N,r,c)

Return N points distributed (roughly) in a uniform manner on the sphere of radius r centered at c.

Inti.geometric_dimension — Function

geometric_dimension(x)

NNumber of degrees of freedom necessary to locally represent the geometrical object. For example, lines have geometric dimension of 1 (whether in ℝ² or in ℝ³), while surfaces have geometric dimension of 2.

Inti.gmsh_curve — Function

gmsh_curve(f::Function, a, b; npts=100, tag=-1)

Create a curve in the current gmsh model given by {f(t) : t ∈ (a,b) } where f is a function from ℝ to ℝ^3. The curve is approximated by C² b-splines passing through npts equispaced in parameter space.

Inti.image — Function

image(f)

Given a function-like object f: Ω → R, return f(Ω).

Inti.import_mesh_from_gmsh_file — Function

import_mesh_from_gmsh_file(fname::String; dim=3)

Open fname and create a LagrangeMesh from the gmsh model in it.

See also: import_mesh_from_gmsh_model.

Inti.import_mesh_from_gmsh_model — Function

import_mesh_from_gmsh_model(;[dim=3]) --> Ω, msh

Create a Domain and a LagrangeMesh from the current gmsh model. Passing dim=2 will create a two-dimensional mesh by projecting the original mesh onto the x,y plane.

This function assumes that the Gmsh API has been initialized through gmsh.initialize.

Inti.integrate — Method

integrate(f,quad::Quadrature)

Compute ∑ᵢ f(qᵢ)wᵢ, where the qᵢ are the quadrature nodes of quad, and wᵢ are the quadrature weights.

Note that you must define f(::QuadratureNode): use q.coords and q.normal if you need to access the coordinate or normal vector at que quadrature node.

Inti.integrate — Method

integrate(f,q::ReferenceQuadrature)
integrate(f,x,w)

Integrate the function f using the quadrature rule q. This is simply sum(f.(x) .* w), where x and w are the quadrature nodes and weights, respectively.

The function f should take an SVector as input.

Inti.interface_method — Method

interface_method(x)

A method of an abstract type for which concrete subtypes are expected to provide an implementation.

Inti.internal_boundary — Method

internal_boundary(Ω::Domain)

Return the internal boundaries of a Domain. These are entities in skeleton(Ω) which appear at least twice as a boundary of entities in Ω.

Inti.jacobian — Method

jacobian(f,x)

Given a (possibly vector-valued) functor f : 𝐑ᵐ → 𝐅ⁿ, return the n × m matrix Aᵢⱼ = ∂fᵢ/∂xⱼ.Both x and f(x) are expected to be of SVector type.

Inti.lagrange_basis — Method

lagrange_basis(E::Type{<:LagrangeElement})

Return the Lagrange basis B for the element E. Evaluating B(x) yields the value of each basis function at x.

Inti.new_tag — Method

new_tag(dim)

Return a new tag for an entity of dimension dim so that EntityKey(dim, tag) is not already in ENTITIES.

Inti.nodes — Method

nodes(msh::SubMesh)

A view of the nodes of the parent mesh belonging to the submesh. The ordering is given by the nodetags function.

Inti.nodetags — Method

nodetags(msh::SubMesh)

Return the tags of the nodes in the parent mesh belonging to the submesh.

Inti.normal — Method

normal(q)

Return the normal vector of q, if it exists.

Inti.notimplemented — Method

notimplemented()

Things which should probably be implemented at some point.

Inti.order — Method

order(q::ReferenceQuadrature)

A quadrature of order p (sometimes called degree of precision) integrates all polynomials of degree ≤ p but not ≤ p + 1.

Inti.order — Method

order(el::LagrangeElement)

The order of the element's interpolating polynomial (e.g. a LagrangeLine with 2 nodes defines a linear polynomial, and thus has order 1).

Inti.pde — Method

pde(K::AbstractPDEKernel)

Return the underlying AbstractPDE when K correspond to the kernel related to the underlying Greens function of a PDE.

Inti.polynomial_solutions_vdim — Method

polynomial_solutions_vdim(pde, order)

For every monomial term pₙ of degree order, compute a polynomial Pₙ such that ℒ[Pₙ] = pₙ, where ℒ is the differential operator associated with pde. This function returns {pₙ,Pₙ,γ₁Pₙ}, where γ₁Pₙ is the generalized Neumann trace of Pₙ.

Inti.qcoords — Method

qcoords(q)

Return the coordinate of the quadrature nodes associated with q.

Inti.quadrature_to_node_vals — Method

quadrature_to_node_vals(Q::Quadrature, qvals::AbstractVector)

Given a vector qvals of scalar values at the quadrature nodes of Q, return a vector ivals of scalar values at the interpolation nodes of Q.mesh.

Inti.qweights — Method

qweights(q)

Return the quadrature weights associated with q.

Inti.reference_nodes — Method

reference_nodes(el::LagrangeElement)
reference_nodes(::Type{<:LagrangeElement})

Return the reference nodes on domain(el) used for the polynomial interpolation. The function values on these nodes completely determines the interpolating polynomial.

We use the same convention as gmsh for defining the reference nodes and their order (see node ordering on gmsh documentation).

Inti.return_type — Method

return_type(f[,args...])

The type returned by f(args...), where args is a tuple of types. Falls back to Base.promote_op by default.

A functors of type T with a knonw return type should extend return_type(::T,args...) to avoid relying on promote_op.

Inti.single_double_layer — Method

single_double_layer(; pde, target, source::Quadrature, compression,
correction, derivative = false)

Construct a discrete approximation to the single- and double-layer integral operators for pde, mapping values defined on the quadrature nodes of source to values defined on the nodes of target. If derivative = true, return instead the adjoint double-layer and hypersingular operators (which are the derivative of the single- and double-layer, respectively).

You must choose a compression method and a correction method, as described below.

Compression

The compression argument is a named tuple with a method field followed by method-specific fields. It specifies how the dense linear operators should be compressed. The available options are:

(method = :none, ): no compression is performed, the resulting matrices are dense.

(method =:hmatrix, tol): the resulting operators are compressed using hierarchical matrices with an absolute tolerance tol (defaults to 1e-8).

(method = :fmm, tol): the resulting operators are compressed using the fast multipole method with an absolute tolerance tol (defaults to 1e-8).

Correction

The correction argument is a named tuple with a method field followed by method-specific fields. It specifies how the singular and nearly-singular integrals should be computed. The available options are:

(method = :none, ): no correction is performed. This is not recommented, as the resulting approximation will be inaccurate if the source and target are not sufficiently far apart.

(method = :dim, maxdist): use the density interpolation method to compute the correction. maxdist specifies the distance between source and target points above which no correction is performed (defaults to Inf).

Inti.single_double_layer_potential — Method

single_double_layer_potential(; pde, source)

Return the single- and double-layer potentials for pde as IntegralPotentials.

Inti.skeleton — Method

skeleton(Ω::Domain)

Return all the boundaries of the domain, i.e. the domain's skeleton.

Inti.standard_basis_vector — Method

standard_basis_vector(k, ::Val{N})

Create an SVector of length N with a 1 in the kth position and zeros elsewhere.

Inti.svector — Method

svector(f,n)

Create an SVector of length n, computing each element as f(i), where i is the index of the element.

Inti.uniform_points_circle — Method

uniform_points_circle(N,r,c)

Return N points uniformly distributed on a circle of radius r centered at c.

Inti.vdim_correction — Method

vdim_correction(pde,X,Y,Γ,S,D,V[;interpolation_order, multiplier])

IntiGmshExt._ent_to_mesh! — Method

_ent_to_mesh!(etype2mat, ent2etags, key, shift)

For each element type used to mesh the entity with key:

push into etype2mat::Dict the pair etype=>connectivity;

push into ent2etags::Dict the pair etype=>etags;

where:

etype::DataType determines the type of the element (see _type_tag_to_etype);

connectivity::Matrix{Int} gives the connectity matrix of the elements of type etype;

etags::Vector{Int} gives the tags of the elements of type etype used to mesh the entity with the given key.

IntiGmshExt._etype_to_type_tag — Method

_etype_to_type_tag(E::DataType)

The inverse of _type_tag_to_etype.

IntiGmshExt._import_mesh! — Method

_import_mesh!(msh)

Import the mesh from gmsh into msh as a LagrangeMesh.

IntiGmshExt._type_tag_to_etype — Method

_type_tag_to_etype(tag)

Mapping of gmsh element types, encoded as an integer, to the internal equivalent of those.