Compression methods

Inti.jl wraps several external libraries providing acceleration routines for integral operators. All compression methods in Inti.jl take an IntegralOperator and return a new object that represents a compressed version of the operator. The following methods are available:

assemble_matrix - Assemble the operator as a dense matrix. Not really a compression method, but useful for debugging and small problems.

assemble_hmatrix - Assemble the operator as a hierarchical matrix.

assemble_fmm - Assemble the operator using the fast multipole method.

All of the compression methods take an IntegralOperator and return a new object that represents a compressed version of the operator. Not all methods work with all operators; the next sections provide more details on their usage and limitations.

Dense matrix

Inti.assemble_matrix — Function

assemble_matrix(iop::IntegralOperator; threads = true)

Assemble a dense matrix representation of an IntegralOperator.

Hierarchical matrix

Inti.assemble_hmatrix — Function

assemble_hmatrix(iop[; atol, rank, rtol, eta])

Assemble an H-matrix representation of the discretized integral operator iop using the HMatrices.jl library.

See the documentation of HMatrices for more details on usage and other keyword arguments.

Fast multipole method

Inti.assemble_fmm — Function

assemble_fmm(iop; atol)

Set up a 2D or 3D FMM for evaluating the discretized integral operator iop associated with the pde. In 2D the FMM2D or FMMLIB2D library is used (whichever was most recently loaded) while in 3D FMM3D is used.

FMMLIB2D

FMMLIB2D does no checking for if the targets and sources coincide, and will return Inf values if iop.target !== iop.source, but there is a point x ∈ iop.target such that x ∈ iop.source.